# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: TokenPak Proxy API
  version: "1.0.2"
  description: |
    TokenPak is a drop-in HTTP proxy that sits between your app and LLM APIs.
    It compresses requests, deduplicates cache hits, and tracks token spend —
    with zero SDK changes required.

    ## Usage

    Point your LLM client at `http://localhost:8766` instead of the provider endpoint:

    ```bash
    ANTHROPIC_BASE_URL=http://localhost:8766 claude-code ...
    OPENAI_BASE_URL=http://localhost:8766/v1 openai ...
    ```

    ## Authentication

    The proxy forwards your API key to the upstream provider. Set the standard
    provider environment variables (`ANTHROPIC_API_KEY`, `OPENAI_API_KEY`, etc.)
    or pass them in the `Authorization` / `x-api-key` header as usual.

  contact:
    name: TokenPak
    url: https://github.com/tokenpak/tokenpak
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8766
    description: Local proxy (default port)
...
paths:
...
  /ingest:
    post:
      operationId: ingestDocument
      summary: Ingest a document into vault context
      description: Parses and indexes a document (text, markdown, code, etc.) into the vault for context injection.
      tags: [ingest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
            example:
              content: "# My Document\nThis is the content to index."
              source_path: /home/user/docs/my-doc.md
              risk_class: low
      responses:
        "200":
          description: Ingest result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /ingest/batch:
    post:
      operationId: ingestBatch
      summary: Batch ingest multiple documents
      description: Ingest multiple documents in a single request.
      tags: [ingest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBatchRequest"
      responses:
        "200":
          description: Batch ingest results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ─── Interactive Docs ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    IngestRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: Document text to ingest
        source_path:
          type: string
          description: Logical path for this document (used as block ID)
        risk_class:
          type: string
          enum: [low, medium, high]
          default: low
    IngestResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        block_id:
          type: string
        tokens:
          type: integer
    IngestBatchRequest:
      type: object
      required: [documents]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/IngestRequest"
    IngestBatchResponse:
      type: object
      properties:
        status:
          type: string
        ingested:
          type: integer
        errors:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/IngestResponse"

    # ── Errors ────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              example: not_found
            message:
              type: string
              example: No trace found for request_id
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"